refacto: move keyboard interactions to custom hook
